% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.ratio.glm.R
\name{calc.oddsratio.glm}
\alias{calc.oddsratio.glm}
\title{Calculate odds ratio of GLM(M)}
\usage{
calc.oddsratio.glm <- function(model, data, incr, quietly = FALSE)
}
\arguments{
\item{data}{The data used for model fitting}

\item{model}{A fitted GLM(M)}

\item{incr}{List. Increment values of each predictor}

\item{quietly}{Logical. Default = FALSE. Whether to output information to the console.}
}
\description{
Function to calculate odds ratio for specific increment steps of GLM.
}
\details{
Currently supported functions: 'stats::glm', 'mgcv::glmmPQL'
}
\examples{
## Example with stats::glm()
# load data (Data source: http://www.ats.ucla.edu/stat/r/dae/logit.htm)
dat <- read.csv("http://www.ats.ucla.edu/stat/data/binary.csv")
dat$rank <- factor(dat$rank)
fit.glm <- glm(admit ~ gre + gpa + rank, data = dat, family = "binomial") # fit model

# Calculate OR for specific increment step of continuous variable
calc.oddsratio.glm(fit.glm, dat, list(gre = 380, gpa = 5))

## Example with MASS:glmmPQL()
# load data
library(nlme)
library(MASS)
data(bacteria)
fit.glmmPQL <- MASS::glmmPQL(y ~ trt + week, random = ~1 | ID,
                             family = binomial, data = bacteria, verbose = FALSE)

# Apply function
calc.oddsratio.glm(fit.glmmPQL, bacteria, list(week = 5))

}

