% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.gam}
\alias{data.gam}
\title{data.gam}
\format{a \code{data.frame} randomly created numerical and non-numerical variables}
\source{
Taken from ?mgcv::gam and added variable "x4"
}
\usage{
data(data.gam)
}
\description{
data.gam
}
\keyword{datasets}
\keyword{internal}

