% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_gam}
\alias{data_gam}
\title{data_gam}
\format{
A \link{data.frame} of randomly created numerical and non-numerical
variables.
}
\source{
Taken from ?mgcv::gam and added variable "x4"
}
\usage{
data(data_gam)
}
\description{
data_gam
}
\keyword{datasets}
\keyword{internal}
