% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_glm}
\alias{data_glm}
\title{data_glm}
\format{a \code{data.frame} randomly created numerical and non-numerical variables}
\source{
Taken from \url{http://www.ats.ucla.edu/stat/r/dae/logit.htm}, direct download
link: \url{http://www.ats.ucla.edu/stat/data/binary.csv}
}
\usage{
data(data_glm)
}
\description{
data_glm
}
\keyword{datasets}
\keyword{internal}
