% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smooth.gam.R
\name{plot_smooth.gam}
\alias{plot_smooth.gam}
\title{Plot smoothing functions of GAM(M) models}
\usage{
plot_smooth.gam(model, pred, col.line = "blue", col.line.ci = "black",
  line.type.ci = "dashed", fill.ci = "grey", title = NULL, xlab = NULL,
  ylab = NULL)
}
\arguments{
\item{model}{A fitted model of class \code{gam}.}

\item{pred}{The predictor of the fitted model to plot the smooth function of.}

\item{col.line}{Character. Sets color for smoothing function. Default to \code{"blue"}}

\item{col.line.ci}{Character. Sets color for confident interval line of smoothing function. Default to \code{"black"}}

\item{line.type.ci}{Character. Sets linetype of confident interval line of smoothing function. Default to \code{"dashed"}.}

\item{fill.ci}{Character. Fill color of area between smoothing function and its confident interval lines}

\item{title}{Character. Plot title.}

\item{xlab}{Character. X-axis title.}

\item{ylab}{Character. Y-axis title.}
}
\description{
This function plots the smoothing function of selected GAM(M) models
using the \code{ggplot2} plotting system.
}
\examples{
suppressPackageStartupMessages(library(mgcv))
set.seed(1234)
n <- 200
sig <- 2
dat <- gamSim(1, n = n, scale = sig, verbose = FALSE)
dat$x4 <- as.factor(c(rep("A", 50), rep("B", 50), rep("C", 50), rep("D", 50)))
fit.gam <- mgcv::gam(y ~ s(x0) + s(I(x1^2)) + s(x2) + offset(x3) + x4, data = dat)

library(oddsratio)
plot_smooth.gam(fit.gam, pred = "x2", title = "Predictor 'x2'")

}
\author{
Patrick Schratz <patrick.schratz@gmail.com>
}
\seealso{
\code{\link[oddsratio]{plot_smooth.gam}}

\code{\link[oddsratio]{calc.oddsratio.gam}}

\code{\link[oddsratio]{add.oddsratio.into.plot}}
}

