% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.oddsratio.into.plot.R
\name{add.oddsratio.into.plot}
\alias{add.oddsratio.into.plot}
\title{Insert odds ratios of GAM(M)s into smoothing function}
\usage{
add.oddsratio.into.plot(plot.object, or.object, col.line = "red",
  col.text = "red", values = TRUE, height.or = 0, height.val = 0,
  x.shift = NULL)
}
\arguments{
\item{plot.object}{A `ggplot` object from \code{\link[oddsratio]{plot_smooth.gam}}}

\item{or.object}{A returned data.frame from \code{\link[oddsratio]{calc.oddsratio.gam}}}

\item{col.line}{Character. Color of the vertical line showing the predictor values}

\item{col.text}{Character. Color of the inserted odds ratio information}

\item{values}{Logical. Whether to print predictor value information nearby 
the inserted vertical lines. Default to \code{TRUE}.}

\item{height.or}{Numeric. Specifies y-location of inserted odds ratio information.
Relative to plotted y-axis range. A positive/negative value will place the 
odds ratio information higher/lower.}

\item{height.val}{Numeric. Specifies y-location of inserted predictor information.
Relative to plotted y-axis range. A positive/negative value will place the 
predictor information higher/lower.}

\item{x.shift}{Numeric. Placement of predictor information relative to its vertical line.
Relative to x-axis range. Default to \code{0.02}. 
Play around to find the perfect position for you.}
}
\value{
Returns a \code{ggplot} plotting object
}
\description{
This function inserts calculated odds ratios of GAM(M)s into a plot of a
GAM(M) smoothing function.
}
\details{
The logic behind this function is to add calculated odds ratio of 
fitted GAM models (\code{\link[oddsratio]{calc.oddsratio.gam}}) into a plot 
showing the smooth function (\code{\link[oddsratio]{plot_smooth.gam}}) of the chosen 
predictor for which the odds ratio was calculated for. Multiple insertions can 
be made by iteratively calling the function (see examples).

Right now the function does only accept results of 
\code{\link[oddsratio]{calc.oddsratio.gam}} with \code{slice = FALSE}. 
If you want to insert multiple odds ratio you have to do it iteratively.
}
\examples{
# load data (Source: ?mgcv::gam)
library(mgcv)
n <- 200
sig <- 2
dat <- gamSim(1, n = n, scale = sig, verbose = FALSE)
dat$x4 <- as.factor(c(rep("A", 50), rep("B", 50), rep("C", 50), rep("D", 50)))
fit.gam <- gam(y ~ s(x0) + s(I(x1^2)) + s(x2) + 
               offset(x3) + x4, data = dat) # fit model
               
# create input objects (plot + odds ratios)   
library(oddsratio)            
plot.object <- plot_smooth.gam(fit.gam, pred = "x2", title = "Predictor 'x2'")
or.object1 <- calc.oddsratio.gam(data = dat, model = fit.gam, pred = "x2", 
                               values = c(0.099, 0.198))
                               
# insert first odds ratios to plot
plot.object <- add.oddsratio.into.plot(plot.object, or.object1, height.or = 5,
                              x.shift = 0.03)

# calculate second odds ratio
or.object2 <- calc.oddsratio.gam(data = dat, model = fit.gam, pred = "x2", 
                                  values = c(0.4, 0.6))
                                  
# add second or into plot                                  
add.oddsratio.into.plot(plot.object, or.object2, height.or = 3, col.line = "green4",
               col.text = "green4")          
}
\author{
Patrick Schratz <patrick.schratz@gmail.com>
}
\seealso{
\code{\link[oddsratio]{plot_smooth.gam}}

\code{\link[oddsratio]{calc.oddsratio.gam}}
}

